backend/requirements.txt
# ============================================================
# MedAI — Backend Requirements
# Python 3.11
# ============================================================

# --- Web Framework ---
fastapi==0.135.2
uvicorn==0.42.0
starlette==1.0.0
python-multipart==0.0.22

# --- ML / Computer Vision ---
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
keras==3.14.0
numpy==1.26.4
opencv-python==4.13.0.92
pillow==12.2.0
scikit-learn==1.8.0
scipy==1.17.1

# --- RAG Pipeline ---
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.83
langchain-openai==0.3.24
langchain-pinecone==0.2.8
langchain-text-splitters==0.3.11
langsmith==0.7.25
openai==1.109.1
pinecone==7.3.0
pinecone-plugin-interface==0.0.7
sentence-transformers==4.1.0
transformers==5.5.0
tokenizers==0.22.2
tiktoken==0.12.0
torch==2.11.0
huggingface-hub==1.9.0
safetensors==0.7.0

# --- Auth & Security ---
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.3.0
cryptography==46.0.6

# --- Database ---
pymongo==4.16.0
dnspython==2.8.0

# --- Data Validation ---
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1

# --- HTTP / Async ---
httpx==0.28.1
httpcore==1.0.9
requests==2.33.1
aiohttp==3.13.0
anyio==4.13.0

# --- Utilities ---
python-dotenv==1.1.0
pyyaml==6.0.3
tqdm==4.67.3
tenacity==9.1.4
orjson==3.11.8
pypdf==5.6.1